[icon] renamed toolkit to platform
