+++ semcore/icon/transform-svg.config.js
@@ -1,12 +1,12 @@
 const getConfig = () => {
   return {
     tasks: ({ generateIcons, getDescriptionIcons, getDescriptionExternalIcons }) => {
-      const getDescriptionToolkitIcons = (iconPath, outLib) => {
+      const getDescriptionPlatformIcons = (iconPath, outLib) => {
         const name = iconPath.split('/').pop().replace('.svg', '');
         return {
           name,
           location: `${outLib}/${name}/index.js`,
-          group: 'toolkit',
+          group: 'platform',
           type: null
         };
       };
@@ -15,7 +15,7 @@ const getConfig = () => {
         generateIcons('svg/color', './color', getDescriptionIcons),
         generateIcons('svg/external', './external', getDescriptionExternalIcons),
         generateIcons('svg/pay', './pay', getDescriptionIcons),
-        generateIcons('svg/toolkit', './toolkit', getDescriptionToolkitIcons),
+        generateIcons('svg/platform', './platform', getDescriptionPlatformIcons),
         generateIcons('svg/icon', '.', getDescriptionIcons),
       ];
     },
